// web/src/api/generated/model/book.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * TCFL Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { ReservationOut } from './reservationOut';

export interface Book {
  id: string;
  title: string;
  authors: string[];
  edition: string | null;
  year: number | null;
  call_number: string | null;
  e_resource_url: string | null;
  copies: number;
  available: number;
  on_loan_to_you: boolean;
  reservation: ReservationOut | null;
  note?: string | null;
}